requirements.txt
google-cloud-aiplatform
feedparser~=6.0.11
requests~=2.32.3
readability-lxml~=0.8.1
beautifulsoup4~=4.12.3
lxml_html_clean
openai~=1.58.1

protobuf~=5.29.2
Jinja2~=3.1.5
boto3~=1.35.90
botocore~=1.35.90
python-dateutil~=2.9.0.post0

python-dotenv~=1.0.1